requirements.txt
alpaca-py>=0.38.0
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3
python-dotenv>=1.0.0
